Frame Driven Audio Sync Attempt
Due to the players input being handled on the frame update, I attempted to play music offset by the frame delta from the quantized marker. This path lacks the needed fidelity of this system, some I'm reorienting to a solution that involves a quartz clock and unmuting samples that are already playing.